Add files via upload
Установлен Redis.
Установлен Celery.
Произведены необходимые конфигурации Django для соединения всех компонент системы.
Реализована рассылка уведомлений подписчикам после создания новости.
Реализована еженедельная рассылка с последними новостями (каждый понедельник в 8:00 утра).
