+++ NewsPaper/celerybeat-schedule.dir
@@ -1,0 +1,4 @@
+'entries', (0, 325)
+'__version__', (512, 15)
+'tz', (1024, 13)
+'utc_enabled', (1536, 4)
