+++ NewsPaper/news/management/commands/runapscheduler.py
@@ -1,90 +1,90 @@
-import datetime
-import logging
-
-from apscheduler.schedulers.blocking import BlockingScheduler
-from apscheduler.triggers.cron import CronTrigger
-from django.conf import settings
-from django.core.mail import EmailMultiAlternatives
-from django.core.management.base import BaseCommand
-from django.template.loader import render_to_string
-from django_apscheduler.jobstores import DjangoJobStore
-from django_apscheduler.models import DjangoJobExecution
-
-from news.models import *
-
-logger = logging.getLogger(__name__)
-
-
-def my_job():
-    today = datetime.datetime.now()
-    last_week = today - datetime.timedelta(days=7)
-    posts = Post.objects.filter(dateCreation__gte=last_week)
-    categories = set(posts.values_list('postCategory__name', flat=True))
-    subscriptions = (Category.objects.filter(name__in=categories).values_list('subscriptions__user__email', flat=True))
-    html_content = render_to_string(
-        'daily_post.html',
-        {
-            'link': settings.SITE_URL,
-            'posts': posts,
-        }
-    )
-    msg = EmailMultiAlternatives(
-        subject='Статьи за неделю',
-        body='',
-        from_email=settings.DEFAULT_FROM_EMAIL,
-        to=subscriptions,
-    )
-
-    msg.attach_alternative(html_content, 'text/html')
-    msg.send()
-
-
-def delete_old_job_executions(max_age=604_800):
-    """
-    This job deletes APScheduler job execution entries older than `max_age`
-    from the database.
-    It helps to prevent the database from filling up with old historical
-    records that are no longer useful.
-
-    :param max_age: The maximum length of time to retain historical
-                    job execution records. Defaults to 7 days.
-    """
-    DjangoJobExecution.objects.delete_old_job_executions(max_age)
-
-
-class Command(BaseCommand):
-    help = "Runs APScheduler."
-
-    def handle(self, *args, **options):
-        scheduler = BlockingScheduler(timezone=settings.TIME_ZONE)
-        scheduler.add_jobstore(DjangoJobStore(), "default")
-
-        scheduler.add_job(
-            my_job,
-            trigger=CronTrigger(
-                day_of_week="fri", hour="18", minute="00"
-            ),
-            id="my_job",
-            max_instances=1,
-            replace_existing=True,
-        )
-        logger.info("Added job 'my_job'.")
-
-        scheduler.add_job(
-            delete_old_job_executions,
-            trigger=CronTrigger(
-                day_of_week="mon", hour="00", minute="00"
-            ),
-            id="delete_old_job_executions",
-            max_instances=1,
-            replace_existing=True,
-        )
-        logger.info("Added weekly job: 'delete_old_job_executions'.")
-
-        try:
-            logger.info("Starting scheduler...")
-            scheduler.start()
-        except KeyboardInterrupt:
-            logger.info("Stopping scheduler...")
-            scheduler.shutdown()
-            logger.info("Scheduler shut down successfully!")
+# import datetime
+# import logging
+#
+# from apscheduler.schedulers.blocking import BlockingScheduler
+# from apscheduler.triggers.cron import CronTrigger
+# from django.conf import settings
+# from django.core.mail import EmailMultiAlternatives
+# from django.core.management.base import BaseCommand
+# from django.template.loader import render_to_string
+# from django_apscheduler.jobstores import DjangoJobStore
+# from django_apscheduler.models import DjangoJobExecution
+#
+# from news.models import *
+#
+# logger = logging.getLogger(__name__)
+#
+#
+# def my_job():
+#     today = datetime.datetime.now()
+#     last_week = today - datetime.timedelta(days=7)
+#     posts = Post.objects.filter(dateCreation__gte=last_week)
+#     categories = set(posts.values_list('postCategory__name', flat=True))
+#     subscriptions = (Category.objects.filter(name__in=categories).values_list('subscriptions__user__email', flat=True))
+#     html_content = render_to_string(
+#         'daily_post.html',
+#         {
+#             'link': settings.SITE_URL,
+#             'posts': posts,
+#         }
+#     )
+#     msg = EmailMultiAlternatives(
+#         subject='Статьи за неделю',
+#         body='',
+#         from_email=settings.DEFAULT_FROM_EMAIL,
+#         to=subscriptions,
+#     )
+#
+#     msg.attach_alternative(html_content, 'text/html')
+#     msg.send()
+#
+#
+# def delete_old_job_executions(max_age=604_800):
+#     """
+#     This job deletes APScheduler job execution entries older than `max_age`
+#     from the database.
+#     It helps to prevent the database from filling up with old historical
+#     records that are no longer useful.
+#
+#     :param max_age: The maximum length of time to retain historical
+#                     job execution records. Defaults to 7 days.
+#     """
+#     DjangoJobExecution.objects.delete_old_job_executions(max_age)
+#
+#
+# class Command(BaseCommand):
+#     help = "Runs APScheduler."
+#
+#     def handle(self, *args, **options):
+#         scheduler = BlockingScheduler(timezone=settings.TIME_ZONE)
+#         scheduler.add_jobstore(DjangoJobStore(), "default")
+#
+#         scheduler.add_job(
+#             my_job,
+#             trigger=CronTrigger(
+#                 day_of_week="fri", hour="18", minute="00"
+#             ),
+#             id="my_job",
+#             max_instances=1,
+#             replace_existing=True,
+#         )
+#         logger.info("Added job 'my_job'.")
+#
+#         scheduler.add_job(
+#             delete_old_job_executions,
+#             trigger=CronTrigger(
+#                 day_of_week="mon", hour="00", minute="00"
+#             ),
+#             id="delete_old_job_executions",
+#             max_instances=1,
+#             replace_existing=True,
+#         )
+#         logger.info("Added weekly job: 'delete_old_job_executions'.")
+#
+#         try:
+#             logger.info("Starting scheduler...")
+#             scheduler.start()
+#         except KeyboardInterrupt:
+#             logger.info("Stopping scheduler...")
+#             scheduler.shutdown()
+#             logger.info("Scheduler shut down successfully!")
