+++ NewsPaper/news/signals.py
@@ -5,26 +5,7 @@ from django.dispatch import receiver
 from django.template.loader import render_to_string
 
 from .models import *
-
-
-def send_notifications(preview, pk, title, subscribers_emails):
-    html_content = render_to_string(
-        'post_created_email.html',
-        {
-            'text': preview,
-            'link': f'http://127.0.0.1:8000/news/{pk}'
-        }
-    )
-
-    msg = EmailMultiAlternatives(
-        subject=title,
-        body='',
-        from_email=settings.DEFAULT_FROM_EMAIL,
-        to=subscribers_emails,
-    )
-
-    msg.attach_alternative(html_content, 'text/html')
-    msg.send()
+from .tasks import send_notifications
 
 
 @receiver(m2m_changed, sender=PostCategory)
@@ -38,3 +19,25 @@ def notify_about_new_post(sender, instance, **kwargs):
             subscribers_emails += [subs.user.email for subs in subscribers]
 
         send_notifications(instance.preview(), instance.pk, instance.title, subscribers_emails)
+
+
+# def send_notifications(preview, pk, title, subscribers_emails):
+#     html_content = render_to_string(
+#         'post_created_email.html',
+#         {
+#             'text': preview,
+#             'link': f'http://127.0.0.1:8000/news/{pk}'
+#         }
+#     )
+#
+#     msg = EmailMultiAlternatives(
+#         subject=title,
+#         body='',
+#         from_email=settings.DEFAULT_FROM_EMAIL,
+#         to=subscribers_emails,
+#     )
+#
+#     msg.attach_alternative(html_content, 'text/html')
+#     msg.send()
+
+
