+++ NewsPaper/news/tasks.py
@@ -1,0 +1,53 @@
+import datetime
+
+from celery import shared_task
+from django.conf import settings
+from django.core.mail import EmailMultiAlternatives
+from django.template.loader import render_to_string
+
+from news.models import Post, Category
+
+
+@shared_task
+def send_notifications(preview, pk, title, subscribers_emails):
+    html_content = render_to_string(
+        'post_created_email.html',
+        {
+            'text': preview,
+            'link': f'http://127.0.0.1:8000/news/{pk}'
+        }
+    )
+
+    msg = EmailMultiAlternatives(
+        subject=title,
+        body='',
+        from_email=settings.DEFAULT_FROM_EMAIL,
+        to=subscribers_emails,
+    )
+
+    msg.attach_alternative(html_content, 'text/html')
+    msg.send()
+
+@shared_task
+def send_weekly_notifications():
+    today = datetime.datetime.now()
+    last_week = today - datetime.timedelta(days=7)
+    posts = Post.objects.filter(dateCreation__gte=last_week)
+    categories = set(posts.values_list('postCategory__name', flat=True))
+    subscriptions = (Category.objects.filter(name__in=categories).values_list('subscriptions__user__email', flat=True))
+    html_content = render_to_string(
+        'daily_post.html',
+        {
+            'link': settings.SITE_URL,
+            'posts': posts,
+        }
+    )
+    msg = EmailMultiAlternatives(
+        subject='Статьи за неделю',
+        body='',
+        from_email=settings.DEFAULT_FROM_EMAIL,
+        to=subscriptions,
+    )
+
+    msg.attach_alternative(html_content, 'text/html')
+    msg.send()
